+++ src/main/java/wolox/albumes/dtos/PostDTO.java
@@ -1,24 +1,24 @@
 package wolox.albumes.dtos;
 
 public class PostDTO {
-    private Integer userId;
-    private Integer id;
+    private Long userId;
+    private Long id;
     private String title;
     private String body;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
